% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceldt.R
\docType{data}
\name{accel}
\alias{accel}
\title{Acceleration Data}
\format{
A data frame with x rows and y columns.
}
\source{
Describe the source of the data
}
\usage{
accel
}
\description{
A dataset containing acceleration measurements.
}
\keyword{datasets}
