% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adae}
\alias{adae}
\title{ADAE table of Panitumumab}
\format{
\subsection{\code{adae}}{

A data frame with 3,211 rows and 19 columns:
\describe{
\item{SUBJID}{Subject ID}
\item{AESEVCD}{Grade/Severity Code}
}
}
}
\usage{
adae
}
\description{
ADAE Panitumumab data for BIS620
}
\keyword{datasets}
