% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_miss_summary_stat.R
\name{add_miss_summary_stat}
\alias{add_miss_summary_stat}
\title{Add the Missing Columns to the Summary Table}
\usage{
add_miss_summary_stat(x, col, x_axis, na_fill = "NA")
}
\arguments{
\item{x}{A tibble data.}

\item{col}{The column we want to add additional row to.}

\item{x_axis}{The value that the \code{x$col} may take.}

\item{na_fill}{The value to fill NA value. (default "NA")}
}
\value{
A tibble data after adding the missing columns.
}
\description{
This function add the values in \code{x_axis} that is not in \code{x$col} to \code{x$col} with their count equal to zero.
}
