% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{ADSL table of Panitumumab}
\format{
\subsection{\code{adsl}}{

A data frame with 935 rows and 7 columns:
\describe{
\item{SUBJID}{Subject ID}
\item{TRT}{Assigned Treatment}
\item{ATRT}{Actual Treatment}
\item{AGE}{Age in Years at Screening}
\item{SEX}{Sex}
}
}
}
\usage{
adsl
}
\description{
ADSL Panitumumab data for BIS620
}
\keyword{datasets}
