% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events_by_race_top_10.R
\name{adverse_events_by_race_top_10}
\alias{adverse_events_by_race_top_10}
\title{Adverse Events by Race (Top 10)}
\usage{
adverse_events_by_race_top_10(data)
}
\arguments{
\item{data}{A data frame containing adverse events and race information.}
}
\value{
A bar plot visualizing the distribution of top 10 adverse events by race.
}
\description{
This function generates a bar plot showing the distribution of top 10 adverse events
by race.
}
\examples{
adverse_events_by_race_top_10(my_data)

}
