% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Age-Based Survival Analysis}
\usage{
age(dl)
}
\arguments{
\item{dl}{A list containing data frames necessary for creating biodata.}
}
\value{
The function does not return a value but generates multiple survival plots
and prints the results of survival difference analyses for each group.
}
\description{
This function performs survival analysis based on age groups within different biomarker groups.
It generates survival plots and performs survival difference analysis for mid-age and old-age groups,
as well as for mutant and wild-type biomarker groups.
}
\examples{
data(NCT00364013, package = "bis620.2023")
dl <- NCT00364013
age(dl)

}
