% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar_plot}
\alias{bar_plot}
\title{A bar plot with the number of counts per group}
\usage{
bar_plot(data, x_var, y_var)
}
\arguments{
\item{data}{the data to create bar plot for}

\item{x_var}{the x variable with different groups}

\item{y_var}{the number of counts per group}
}
\value{
bar plot of number of subjects for different group
}
\description{
A bar plot with the number of counts per group
}
