% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_long.R
\name{bio_long}
\alias{bio_long}
\title{A function to get pivoted biomark data}
\usage{
bio_long(data)
}
\arguments{
\item{data}{the cleaned biomark data}
}
\value{
a pivoted biomark data indicating number of mutant, wild type, and failure
}
\description{
A function to get pivoted biomark data
}
