% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomark}
\alias{biomark}
\title{Biomark table of Panitumumab}
\format{
\subsection{\code{biomark}}{

A data frame with 935 rows and 19 columns:
\describe{
\item{SUBJID}{Subject ID}
\item{TRT}{Assigned Treatment}
\item{BMMTR1, BMMTR2}{exon Result 1}
}
}
}
\usage{
biomark
}
\description{
Biomark Panitumumab data for BIS620
}
\keyword{datasets}
