% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brief_summaries}
\alias{brief_summaries}
\title{brief_summaries data in ctgov.duckdb}
\format{
\subsection{\code{brief_summaries}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{nct_id}{id of nct}
\item{X, Y, Z}{Accelerometry measurement (in milligravities).}
}
}
}
\usage{
brief_summaries
}
\description{
Toy data for BIS620
}
\keyword{datasets}
