% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_duration.R
\name{calculate_duration}
\alias{calculate_duration}
\title{duration calculation}
\usage{
calculate_duration(start_date, end_date, unit = "days")
}
\arguments{
\item{start_date}{start_date of the trail}

\item{end_date}{completion date of the trail}

\item{unit}{end unit which can days, months, years}
}
\description{
This function can calculate the duration between start_date and end_date in days
and can transform duration into months and years.
}
