% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calculated_values}
\alias{calculated_values}
\title{calculated_values Data from ctgov}
\format{
\subsection{\code{calculated_values}}{

A tibble with various information about trials
}
}
\usage{
calculated_values
}
\description{
calculated_values data for HW3
}
\keyword{datasets}
