% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_environment.R
\name{check_environment}
\alias{check_environment}
\title{Create the connection to a database and "studies" and "sponsors" tables.}
\usage{
check_environment()
}
\value{
A database connection object.
}
\description{
Checks the local environment and establishes a database connection.
}
