% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_sq.R
\name{chi_sq}
\alias{chi_sq}
\title{Chi-squared test for significance between variables}
\usage{
chi_sq(var1, var2)
}
\arguments{
\item{var1}{the first categorical variable to test}

\item{var2}{the second categorical variable to test}
}
\value{
a text indicating whether the correlation is significant
}
\description{
Chi-squared test for significance between variables
}
