% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_analysis.R
\name{chisq_analysis}
\alias{chisq_analysis}
\title{Chi-squared Analysis}
\usage{
chisq_analysis(data, column1, column2)
}
\arguments{
\item{data}{A data frame containing the variables to analyze.}

\item{column1}{The name of the first categorical variable.}

\item{column2}{The name of the second categorical variable.}
}
\value{
The chi-squared test result and the contingency table.
}
\description{
This function performs a chi-squared test of independence between two categorical variables.
}
\examples{
chisq_analysis(my_data, "Variable1", "Variable2")

}
