% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_bio.R
\name{clean_biomark}
\alias{clean_biomark}
\title{A function to clean biomark data}
\usage{
clean_biomark(data)
}
\arguments{
\item{data}{the biomark data}
}
\value{
a cleaned biomark data indicating number of mutant, wild type, and failure
}
\description{
A function to clean biomark data
}
