% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_shinyapp.R
\name{clinical_shinyapp}
\alias{clinical_shinyapp}
\title{Start Shiny App for Clinical Trial Data}
\usage{
clinical_shinyapp()
}
\value{
No return.
}
\description{
This function can start our clinical data visualization app.
}
