% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Conditions}
\format{
\subsection{\code{conditions}}{

A data frame with 822,729 rows and 4 columns:
\describe{
\item{id}{A unique identifier}
\item{nct_id}{The NCT id, another identifier}
\item{name}{The name of the condition}
\item{downcase_name}{The same as name but all lowercase}
}
}
}
\usage{
conditions
}
\description{
Taken from ctgov database table, lists conditions
}
\keyword{datasets}
