% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{Correlation plot of the variables in data}
\usage{
cor_plot(data)
}
\arguments{
\item{data}{the data to create correlation plot for}
}
\value{
a correlation plot with numeric variables in data
}
\description{
Correlation plot of the variables in data
}
