% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_id.R
\name{count_country_id}
\alias{count_country_id}
\title{A function for selecting ID count of individual countries}
\usage{
count_country_id(country, countries)
}
\arguments{
\item{country}{the country selected to count number of ID}

\item{countries}{the data frame with all countries and IDs}
}
\value{
the number of ID counts given a specific country
}
\description{
A function for selecting ID count of individual countries
}
