% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{counties Data from social vulnerability data}
\format{
\subsection{\code{counties}}{

A special sf dataframe which contains counties information and geographical
coordinates of it stored in a geometry format
}
}
\usage{
counties
}
\description{
processed data for the final project
}
\keyword{datasets}
