% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Countries}
\format{
\subsection{\code{countries}}{

A data frame with 658,829 rows and 4 columns:
\describe{
\item{id}{A unique identifier}
\item{nct_id}{The NCT id, another identifier}
\item{name}{The name of the country}
\item{removed}{True or False}
}
}
}
\usage{
countries
}
\description{
Taken from ctgov database table, gives the country for each trial
}
\keyword{datasets}
