% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_list}
\alias{country_list}
\title{country_list data in ctgov.duckdb}
\format{
\subsection{\code{country_list}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{nct_id}{id of nct}
\item{X, Y, Z}{Accelerometry measurement (in milligravities).}
}
}
}
\usage{
country_list
}
\description{
Toy data for BIS620
}
\keyword{datasets}
