% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.R
\name{cox}
\alias{cox}
\title{Fit a Cox Proportional Hazards Model}
\usage{
cox(dl)
}
\arguments{
\item{dl}{A list containing data frames necessary for creating biodata.}
}
\value{
The function returns a Cox proportional hazards model object fitted to the data.
}
\description{
This function fits a Cox proportional hazards model to the biodata. It models
the survival time based on several covariates including sex, race, BMI, age, treatment,
and KRAS status. This is a common model used in survival analysis to explore the
relationship between the survival time and one or more predictor variables.
}
\examples{
data(NCT00364013, package = "bis620.2023")
dl <- NCT00364013
cox_model <- cox(dl)

}
