% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_cloud.R
\name{create_condition_cloud}
\alias{create_condition_cloud}
\title{Conditions Word-cloud}
\usage{
create_condition_cloud(d, freq = 10, max = 25)
}
\arguments{
\item{d}{the table used to plot}

\item{freq}{Minimum Frequency of Conditions}

\item{max}{Maximum Number of Conditions}
}
\value{
a histogram of the interventions
}
\description{
create a word-coud plot of conditions
}
