% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_histogram.R
\name{create_condition_histogram}
\alias{create_condition_histogram}
\title{Condition Histogram}
\usage{
create_condition_histogram(d, n_conditions)
}
\arguments{
\item{d}{the table used to plot}

\item{n_conditions}{the number of conditions that show in the histogram}
}
\value{
a histogram of the conditions
}
\description{
Create a histogram of a specific number of conditions
}
