% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_table.R
\name{create_data_table}
\alias{create_data_table}
\title{Query a studies  for the selected eligibility criteria}
\usage{
create_data_table(d)
}
\arguments{
\item{d}{the database table.}
}
\description{
Shows a table of studies that aligns with the participant type selected
}
\author{
Jonathan Lee
}
