% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_histogram.R
\name{create_design_histogram}
\alias{create_design_histogram}
\title{Design Histogram}
\usage{
create_design_histogram(x, way)
}
\arguments{
\item{x}{the database table.}

\item{way}{the filtering choices of allocation ways}
}
\value{
A histogram for the strategy for assigning interventions
to participants.
}
\description{
Create a histogram of the intervention study models.
}
