% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_endpoint_histogram.R
\name{create_endpoint_histogram}
\alias{create_endpoint_histogram}
\title{Create a histogram of the endpoint}
\usage{
create_endpoint_histogram(studies, endpoints, kw)
}
\arguments{
\item{studies}{the studies database table.}

\item{endpoints}{the endpoints database table.}

\item{kw}{the keyword.}
}
\description{
Create a histogram of the endpoint
}
