% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_intervention_pie_data}
\alias{create_intervention_pie_data}
\title{Pie chart of intervention types}
\usage{
create_intervention_pie_data(
  studies,
  sponsors,
  interventions,
  kw,
  sponsortype,
  dates
)
}
\arguments{
\item{studies}{A tibble. The data table to be filtered}

\item{sponsors}{A tibble. The data table to be joined}

\item{interventions}{A tibble. The data table to be filtered}

\item{kw}{A string. The keyword string entered by the users}

\item{sponsortype}{A string. The sponsor type entered by the users}

\item{dates}{A list of two date values (start and end date)}
}
\value{
A pie chart with filtered entries
}
\description{
Function that creates a pie chart showing the distribution of intervention type that trials in a query are examining,
filtered by dates and sponsors information.
}
