% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_intervention_histogram.R
\name{create_interventions_histogram}
\alias{create_interventions_histogram}
\title{Intervention Histogram}
\usage{
create_interventions_histogram(d)
}
\arguments{
\item{d}{the table used to plot}
}
\value{
a histogram of the interventions
}
\description{
Create a histogram of interventions
}
