% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_phase_hist_plot}
\alias{create_phase_hist_plot}
\title{Histogram of study phases}
\usage{
create_phase_hist_plot(studies, sponsors, kw, dates, color, sponsortype)
}
\arguments{
\item{studies}{A tibble. The data table to be filtered}

\item{sponsors}{A tibble. The data table to be joined}

\item{kw}{A string. The keyword string entered by the users}

\item{dates}{A list of two date values (start and end date)}

\item{color}{A string. The color string chosen by the users}

\item{sponsortype}{A string. The sponsortype entered by the users}
}
\value{
A plot with filtered entries
}
\description{
Filters the \code{studies} data based on multiple queries (i.e., title keywords,
date range, sponsor type) and returns returns a histogram showing the distribution
of study phases
}
