% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phase_histogram_plot.R
\name{create_phase_histogram_plot}
\alias{create_phase_histogram_plot}
\title{Phase Histogram}
\usage{
create_phase_histogram_plot(d)
}
\arguments{
\item{d}{the table used to plot}
}
\value{
a histogram of the phase
}
\description{
create a histogram of the phase
}
