% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_purpose_pie}
\alias{create_purpose_pie}
\title{Pie chart of study purposes}
\usage{
create_purpose_pie(studies, sponsortype, kw, dates)
}
\arguments{
\item{studies}{A tibble. The data table to be filtered}

\item{sponsortype}{A string. The sponsor type entered by the users}

\item{kw}{A string. The keyword string entered by the users}

\item{dates}{A list of two date values (start and end date)}
}
\value{
A pie chart plot
}
\description{
Function that filters the studies data based on multiple queries,
including keyword matching, date range filter, and sponsor type, and returns
a pie chart showing distribution of different primary purposes.
}
