% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_start_year_histogram_plot.R
\name{create_start_year_histogram_plot}
\alias{create_start_year_histogram_plot}
\title{create_start_year_histogram_plot}
\usage{
create_start_year_histogram_plot(
  studies,
  sponsor_type,
  status_type,
  brief_title_kw
)
}
\arguments{
\item{studies}{A dataset of clinical studies, serving as the primary source of study information.}

\item{sponsor_type}{A variable representing the sponsor type to filter by. By default, it considers all sponsor types (sponsor_type = 1), but you can specify a specific sponsor type to focus the analysis.}

\item{status_type}{A variable representing the status type to filter by. By default, it considers all status types (status_type = 1), but you can specify a specific status type to narrow down the analysis.}

\item{brief_title_kw}{A keyword to search for in study titles. The default is an empty string, enabling you to further refine the analysis by focusing on studies with specific keywords in their titles.}
}
\value{
A histogram plot that visualizes the distribution of study start years based on the specified criteria. The plot helps in understanding the historical trend of clinical studies and can reveal patterns related to specific sponsor types or status types.
}
\description{
This function creates a histogram plot displaying the distribution of study start years in clinical studies. It allows for optional filtering based on sponsor type, status type, and a keyword search in study titles, making it a versatile tool for exploring the temporal distribution of studies.
}
\details{
Create a Histogram Plot for Different Start Years
}
\author{
Yi Ren
}
