% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_study_type_histogram_plot.R
\name{create_study_type_histogram_plot}
\alias{create_study_type_histogram_plot}
\title{Create a Histogram Plot for Study Types}
\usage{
create_study_type_histogram_plot(
  studies,
  sponsor_type,
  status_type,
  brief_title_kw
)
}
\arguments{
\item{studies}{A data frame containing the details of studies.}

\item{sponsor_type}{The type of sponsor for the studies to be included in the plot.}

\item{status_type}{The type of status of the studies to be included in the plot.}

\item{brief_title_kw}{A keyword to filter the studies.}
}
\value{
A ggplot object representing the histogram of the number of studies by study type.
}
\description{
This feature generates the histogram and enables subsetting of the data based on various study types.
There are five distinct study types: Observational, Observational \link{Patient Registry}, Interventional, Expanded access, unknown.
The "unknown" types showed as "NA" in the histogram.
By showing the frequency of each study type, researchers can perform comparative analyses to understand how different types of studies contribute to the field.
For instance, if interventional studies are more common than observational ones, it may reflect a research trend or indicate where more resources are being allocated.
}
\author{
Yiran Liu
}
