% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util.R
\name{create_studytype_histogram}
\alias{create_studytype_histogram}
\title{Function that filters the studies data based on multiple queries,
including keyword matching, date range filter, sponsor type, and
histogram color choice, and returns returns a histogram showing the
distribution of different study types}
\usage{
create_studytype_histogram(studies, sponsortype, kw, dates, color)
}
\arguments{
\item{studies}{A tibble. The data table to be filtered}

\item{sponsortype}{A string. The sponsortype entered by the users}

\item{kw}{A string. The keyword string entered by the users}

\item{dates}{A list of two date values (start and end date)}

\item{color}{A string. The color string chosen by the users}
}
\value{
A plot with filtered entries
}
\description{
Function that filters the studies data based on multiple queries,
including keyword matching, date range filter, sponsor type, and
histogram color choice, and returns returns a histogram showing the
distribution of different study types
}
