% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_word_cloud_plot.R
\name{create_word_cloud_plot}
\alias{create_word_cloud_plot}
\title{Keywords in Brief Titles Word Cloud Plot}
\usage{
create_word_cloud_plot(studies, sponsor_type, status_type, brief_title_kw)
}
\arguments{
\item{studies}{the 'studies' data frame containing information about all clinical trials}

\item{sponsor_type}{the specified sponsor type for studies filtering}

\item{status_type}{the specified status type for studies filtering}

\item{brief_title_kw}{Keyword(s) for filtering studies by their brief titles.
If this parameter is empty, only the first 1000 studies are considered.}
}
\value{
Displays a word cloud plot representing the frequency of words
in the brief titles of the filtered clinical trials.
}
\description{
Generates a word cloud plot based on the brief titles of the filtered
clinical trials, reflecting the frequency of words
after text pre-processing and filtering.
}
\author{
Houmin Xing
}
