% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_world_map.R
\name{create_world_map}
\alias{create_world_map}
\title{Filtered Clinical Trials Distribution World Map}
\usage{
create_world_map(studies, sponsor_type, status_type, brief_title_kw, countries)
}
\arguments{
\item{studies}{the 'studies' data frame containing information about all clinical trials}

\item{sponsor_type}{the specified sponsor type for studies filtering}

\item{status_type}{the specified status type for studies filtering}

\item{brief_title_kw}{Keyword(s) for filtering studies by their brief titles.
If this parameter is empty, only the first 1000 studies are considered.}

\item{countries}{the 'countries' data frame in our ctgov database}
}
\value{
Displays a world map showing the distribution of clinical trial counts (in log10 scale) based on specified filters.
}
\description{
Generates a world map visualization showing the distribution of the clinical
trial counts across different countries based on the specified search criteria.
}
\author{
Houmin Xing
}
