% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-process.R
\name{data_process}
\alias{data_process}
\title{This function does data processing.}
\usage{
data_process(dl)
}
\arguments{
\item{dl}{an object inherited from list.}
}
\value{
full dataset.
}
\description{
This function does data processing.
}
