% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_analysis.R
\name{demographic_analysis}
\alias{demographic_analysis}
\title{Demographic Analysis}
\usage{
demographic_analysis(data)
}
\arguments{
\item{data}{A data frame containing demographic and adverse events information.}
}
\value{
Chi-square test results for sex and adverse events, and race and adverse events.
}
\description{
This function conducts a chi-square test for the association between demographic variables
(sex, race) and adverse events.
}
\examples{
demographic_analysis(my_data)

}
