% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_bmi.R
\name{describe_bmi}
\alias{describe_bmi}
\title{Descriptive Analysis of BMI}
\usage{
describe_bmi(dl)
}
\arguments{
\item{dl}{A list containing at least one data frame named 'adsl'.
The 'adsl' data frame must include columns 'B_WEIGHT' and 'B_HEIGHT' for BMI calculation.}
}
\value{
A named vector containing descriptive statistics for the BMI column, including mean, median, standard deviation, minimum, and maximum.
}
\description{
This function performs a descriptive analysis on the 'BMI' column of the 'adsl' data frame within the provided list object. It calculates and returns basic descriptive statistics for the BMI values.
}
\examples{
dl <- list(adsl = data.frame(SUBJID = 1:5, ATRT = rep("Treatment", 5), 
                            DTHDY = c(100, 150, 200, 250, 300), 
                            DTH = c(0, 1, 0, 1, 0), AGE = c(30, 35, 40, 45, 50), 
                            SEX = c("Male", "Female", "Male", "Female", "Male"), 
                            B_WEIGHT = c(70, 60, 75, 65, 80), 
                            B_HEIGHT = c(175, 160, 180, 165, 170), 
                            RACE = c("Race1", "Race2", "Race3", "Race1", "Race2")))
describe_bmi(dl)
}
