% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_sex.R
\name{describe_sex}
\alias{describe_sex}
\title{Descriptive Analysis of Sex Categories}
\usage{
describe_sex(dl)
}
\arguments{
\item{dl}{A list containing at least one data frame named 'adsl' with a 'SEX' column.}
}
\value{
A data frame with columns 'Sex', 'Count', and 'Proportion', where each row represents a unique sex category from the input data frame. The 'Count' column contains the number of occurrences of each sex category, and the 'Proportion' column contains the proportion of each sex category relative to the total count.
}
\description{
This function performs a descriptive analysis on the 'SEX' column of the 'adsl' data frame within the provided list object. It calculates the count and proportion of each sex category present in the data.
}
\examples{
dl <- list(adsl = data.frame(SEX = c("Male", "Female", "Female", "Male", "Male")))
describe_sex(dl)
}
