% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{documents}
\alias{documents}
\title{NCT documents}
\format{
\subsection{\code{documents}}{

A data frame with 10,449 rows and 6 columns. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
\describe{
\item{id}{the unique number for each study}
\item{nct_id}{nct id of the study}
\item{document_id}{document id of the study}
\item{document_type}{document type of the study}
\item{url}{url of the study}
\item{comment}{comment of the study}
}
}
}
\usage{
documents
}
\description{
NCT documents data for BIS620
}
\keyword{datasets}
