% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_endpoints.R
\docType{data}
\name{endpoints}
\alias{endpoints}
\title{Endpoints dataset from ctrialsgov database}
\format{
\subsection{\code{endpoints}}{

A data frame with 6,042 rows and 2 columns:
\describe{
\item{nct_id}{The unique identifier assigned to each clinical study registered on ClinicalTrials.gov}
\item{endpoint_met}{Result of endpoint met or not}
}
}
}
\usage{
endpoints
}
\description{
Data for BIS620 HW3
}
\keyword{datasets}
