% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_query.R
\name{exclude_query}
\alias{exclude_query}
\title{Query from a database table with rows excluded based on keywords.}
\usage{
exclude_query(d, ex_kwds, column)
}
\arguments{
\item{d}{the database table.}

\item{ex_kwds}{the keywords want to be excluded.}

\item{column}{the column to look for the keywords in.}
}
\description{
The function returns data that exclude rows with some keywords in a specified column.
}
