% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_project_functions.R
\name{feature_extraction}
\alias{feature_extraction}
\title{Create the features from the geometric column for the further analysis}
\usage{
feature_extraction(county_data)
}
\arguments{
\item{county_data}{sf_tibble/sf_data.frame, input sf dataset which must contain special
geometry column}
}
\description{
this function calculates the distances to each US state for each
US county by taking the minimum across distances between the center of some
county and centers of all counties in the state
}
