% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_status.R
\name{filter_status}
\alias{filter_status}
\title{Filter the status returned after applying the filter}
\usage{
filter_status(x, ls_status)
}
\arguments{
\item{x}{the database table.}

\item{ls_status}{the list of included status to look for.}
}
\description{
Filter the status returned after applying the filter
}
