% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_kras.R
\name{find_kras}
\alias{find_kras}
\title{Determine the KRAS Mutation Status}
\usage{
find_kras(bmmtr)
}
\arguments{
\item{bmmtr}{A character vector representing biological measurement markers.
Each element should be one of "Wild-type", "Mutant", "Unknown", or "Failure".}
}
\value{
A character string indicating the KRAS mutation status.
It returns 'Mutant' if any 'Mutant' is present in the vector.
If no 'Mutant' is found, it compares the count of 'Wild-type' with the sum of
'Failure' and 'Unknown'. If 'Wild-type' count is higher, it returns 'Wild-type',
otherwise, it returns 'Unknown'.
}
\description{
This function analyzes a vector representing biological measurement markers (BMMs)
to determine the KRAS mutation status. It categorizes the status into 'Mutant',
'Wild-type', or 'Unknown' based on the frequency of each type in the input vector.
}
\examples{
find_kras(c("Wild-type", "Wild-type", "Unknown"))
find_kras(c("Mutant", "Unknown", "Failure"))
find_kras(c("Wild-type", "Failure", "Failure"))

}
