% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_project_functions.R
\name{forward_selection_linreg}
\alias{forward_selection_linreg}
\title{Forward selection for the linear regression model}
\usage{
forward_selection_linreg(df, X_cols, y_col)
}
\arguments{
\item{df}{data.frame, dataset which has all the data needed for the modelling}

\item{X_cols}{vector of characters, set of X variables to select from}

\item{y_col}{character, y column/target}
}
\description{
this function performs forward selection from the set of variables
specified by user. The user also specifies the y columns and the dataset which contains
both y and X columns. The function returns the dataframe where each row is the iteration
var columns is the optimal variable added at the forward selection iteration
also the dataset shows AIC, r2 and r2_adj of the model after optimal variable
(optimal at the iteration) was added to the model
}
