% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_project_functions.R
\name{geographical_visualizer}
\alias{geographical_visualizer}
\title{Create a geographical map with the geographical distribution of the variable}
\usage{
geographical_visualizer(data, vis_col)
}
\arguments{
\item{data}{sf_tibble/sf_data.frame, input sf dataset which must contain special
geometry column so that the geographical plot can be made}

\item{vis_col}{character, the variable (column of the data) to be analysed}
}
\description{
this function will print the geographical distribution of some
specified variable over the region covered by the data by showing the map
of the region filled with the color corresponding the value of the
specified variable in the certain region
}
