% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_project_final.R
\name{get_all_phases}
\alias{get_all_phases}
\title{Supporting function for problem 1}
\usage{
get_all_phases(studies)
}
\arguments{
\item{studies}{duckDB table, input "studies" table}
}
\description{
Extracts a vector which has all the unique values of \code{phase} in the input
(unfiltered) table \code{studies};
it was build instead of directly listing all the unique possible values of "phase" so that if the
dataset is updated, the has not no be changed;
it returns a character vector \code{phase_all}
}
