% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biodata.R
\name{get_biodata}
\alias{get_biodata}
\title{Extract and Combine Patient Biodata and Biomarkers}
\usage{
get_biodata(dl)
}
\arguments{
\item{dl}{A list containing at least two data frames: \code{adsl} and \code{biomark}.
\code{adsl} should have columns for subject ID (SUBJID), treatment (ATRT), death day (DTHDY),
death indicator (DTH), age (AGE), sex (SEX), baseline weight (B_WEIGHT), baseline height (B_HEIGHT),
and race (RACE). \code{biomark} should have columns for subject ID (SUBJID) and biomarker measurements
(BMMTR1, BMMTR2, BMMTR3, BMMTR15).}
}
\value{
A data frame combining the patient's biodata with their biomarker information,
including calculated BMI and KRAS mutation status.
}
\description{
This function extracts patient biodata from a dataset and combines it with
biomarker data, including the KRAS mutation status determined by the \code{find_kras} function.
It selects specific columns from the biodata and biomarker datasets and then performs
an inner join on the subject identifier (SUBJID).
}
\examples{
# Assume dl is a list with data frames adsl and biomark
data(NCT00364013, package = "bis620.2023")
dl <- NCT00364013
biodata <- get_biodata(dl)

}
