% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomarker.R
\name{get_biomarker}
\alias{get_biomarker}
\title{Extract Biomarker Subsets}
\usage{
get_biomarker(biodata)
}
\arguments{
\item{biodata}{A data frame containing patient biodata with a column 'KRAS' indicating the biomarker status.}
}
\value{
A list containing two data frames: one with rows where KRAS is 'Mutant' and the other where KRAS is 'Wild-type'.
}
\description{
This function separates the biodata into two subsets based on the KRAS biomarker status: 'Mutant' and 'Wild-type'.
}
\examples{
# Assuming biodata is a data frame with a 'KRAS' column
data(NCT00364013, package = "bis620.2023")
dl <- NCT00364013

biomarker_sets <- get_biomarker(get_biodata(dl))

}
