% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util.R
\name{get_concurrent_studies}
\alias{get_concurrent_studies}
\title{Get the number of concurrent studies for each date in a set of studies.}
\usage{
get_concurrent_studies(studies_df)
}
\arguments{
\item{studies_df}{A data frame. The studies to get the number of concurrent
trials for.}
}
\value{
A tibble with a \code{date} column and a \code{count} of the number of concurrent
trials at that date.
}
\description{
This function takes a data frame of studies and returns a tibble with a
\code{date} column and a \code{count} of the number of concurrent trials at that date.
}
\examples{
\dontrun{
studies_df <- data.frame(
    start_date = as.Date(c('2020-01-01', '2020-02-01', '2020-03-01')),
    completion_date = as.Date(c('2020-12-31', '2020-11-30', '2020-10-31')))
get_concurrent_studies(studies_df)
}
}
