% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_hist.R
\name{get_condition_histogram}
\alias{get_condition_histogram}
\title{Get top 8 frequent number of conditions by a brief title keyword search and sponsor type}
\usage{
get_condition_histogram(study, condition)
}
\arguments{
\item{study}{the studies to get the number of studies trials for.}

\item{condition}{the condition table of the trials.}
}
\value{
values of top 8 frequent number of conditions
}
\description{
Get top 8 frequent number of conditions by a brief title keyword search and sponsor type
}
