% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{get_conditions_for_intervention_type}
\alias{get_conditions_for_intervention_type}
\title{Get interventions on condition mapping}
\usage{
get_conditions_for_intervention_type(interventionType)
}
\arguments{
\item{interventionType}{the type of interventions.}
}
\value{
a plot of top 10 conditions on specific intervention type
}
\description{
Get interventions on condition mapping
}
