% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distinct.R
\name{get_distinct}
\alias{get_distinct}
\title{Get a Vector of Distinct Value of a Column in Tibble}
\usage{
get_distinct(d, column, na_fill = "NA")
}
\arguments{
\item{d}{A tibble data.}

\item{column}{A column name. It should be in the columns of \code{d}.}

\item{na_fill}{the value to fill NA value. (default "NA")}
}
\value{
A vector of distinct value of the column in the tibble.
}
\description{
This function returns a vector of the distinct value of a column in a tibble. It also change the NA to a input value specified by users.
}
