% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facilities_wmap.R
\name{get_facilities_wmap}
\alias{get_facilities_wmap}
\title{The world map of the facilities}
\usage{
get_facilities_wmap(facilities)
}
\arguments{
\item{facilities}{the facilities}
}
\value{
A table of the number of facilities joined with world map data
}
\description{
Feature 2: Create a world map of the number of facilities
}
\author{
Anran Jiao
}
