% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genotype.R
\name{get_genotype}
\alias{get_genotype}
\title{Determine the genotype for each subject ID that is passed to the function.}
\usage{
get_genotype(subjid)
}
\arguments{
\item{subjid}{A string containing information about which subject's biomarkers to examine.}
}
\value{
A string of "Mutant", "Wild-type", or "Unknown" representing the subject's genotype.
}
\description{
This function takes a subject ID (\code{subjid}) and finds all of their biomarkers. A patient will be considered
“Mutant” if there is at least one “Mutant” biomarker in KRAS exons 2, 3, 4. Patients will be considered
“Wild-type” if they are not “Mutant” and they have more “Wild-type” markers than “Unknown” or “Failure”.
}
\details{
The function works by first extracting the subject's unique biomarkers (\code{sublist}). It then
runs through the sublist, and determines "Mutant" if there is at least one “Mutant” biomarker, or
"Wild-type"  if they are not “Mutant” and they have more “Wild-type” markers than “Unknown” or “Failure”.
The result is a string representing the determined genotype.
}
