% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{get_outcome_pie_for_intervention}
\alias{get_outcome_pie_for_intervention}
\title{Get outcome pie chart conditioned on intervention}
\usage{
get_outcome_pie_for_intervention(interventionType)
}
\arguments{
\item{interventionType}{the type of interventions.}
}
\value{
a pie chart of number of outcomes on specific intervention type
}
\description{
Get outcome pie chart conditioned on intervention
}
