% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util.R
\name{get_study_conditions}
\alias{get_study_conditions}
\title{Join studies data with study conditions data.}
\usage{
get_study_conditions(studies_df, conditions_df = CONDITIONS)
}
\arguments{
\item{studies_df}{A data frame containing the studies data.}
}
\value{
A data frame containing the joined studies and study conditions data.
}
\description{
This function joins the studies data frame with the study conditions
data frame on 'nct_id'. The resulting data frame is returned.
}
