% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word.R
\name{get_word}
\alias{get_word}
\title{get_word function for Feature 5}
\usage{
get_word(x)
}
\arguments{
\item{x}{the database table}
}
\value{
word counts of the database table
}
\description{
get the word counts from a column of database
}
\references{
https://shiny.posit.co/r/gallery/start-simple/word-cloud/
}
\author{
Min Zhu
}
