% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTermMatrix.R
\name{getTermMatrix}
\alias{getTermMatrix}
\title{Word Frequency}
\usage{
getTermMatrix(x)
}
\arguments{
\item{x}{the text.}
}
\value{
Frequency of words with descending order.
}
\description{
Word Frequency
}
