% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print "hello world"}
\usage{
hello(who = "world")
}
\arguments{
\item{who}{Who to say hello to. (Default is "world")}
}
\description{
Print the hello world function to the console with
parameter way who to way hello to.
}
\examples{
hello("people in BIS620")
}
