% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_interventions.R
\docType{data}
\name{interventions}
\alias{interventions}
\title{Interventions dataset from ctrialsgov database}
\format{
\subsection{\code{interventions}}{

A data frame with 793,532 rows and 2 columns:
\describe{
\item{nct_id}{The unique identifier assigned to each clinical study registered on ClinicalTrials.gov}
\item{intervention_type}{The type of interventions}
}
}
}
\usage{
interventions
}
\description{
Data for BIS620 HW3
}
\keyword{datasets}
