% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lauch_shiny.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Launch shiny app}
\usage{
launch_shiny(path = "../ctrialsgovdb")
}
\arguments{
\item{path}{(string) that defines where the databaset is stored, default ../ctrialsgovdb}
}
\description{
This function launch the shiny app, the function requires that the duckdb ctgov database is localy saved,
this can be done through \link[bis620.2023]{download_clinicaltrials()}.
}
