% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{launchApp}
\alias{launchApp}
\title{Launch My Shiny App}
\usage{
launchApp(con)
}
\arguments{
\item{con}{the connection to the duckdb file that needed to launch}
}
\description{
This function launches the Shiny application clinical studies and trials.
}
