% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_model.R
\name{lm_model}
\alias{lm_model}
\title{Linear model for significance between variables}
\usage{
lm_model(data, expression)
}
\arguments{
\item{data}{the data used for the model}

\item{expression}{expression of the linear model}
}
\value{
a linear regression model
}
\description{
Linear model for significance between variables
}
