% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewFunc1.R
\name{binary.barplot}
\alias{binary.barplot}
\title{Create Bar Plots Grouped by a Binary Variable}
\usage{
binary.barplot(mydata, x.vars, group.var)
}
\arguments{
\item{mydata}{Data frame containing the variables for plotting.}

\item{x.vars}{Vector of strings representing the names of independent variables in 'mydata' to be used for plotting.}

\item{group.var}{String representing the name of the binary dependent variable in 'mydata' by which the bar plot will be grouped and filled.}
}
\value{
Prints a sequence of bar plots. Each plot corresponds to an independent variable in 'x.vars', grouped and filled by the binary variable 'group.var'. No return value.
}
\description{
This function generates and prints a sequence of bar plots for a set of specified independent variables, grouped and filled by a binary dependent variable. Each independent variable specified in 'x.vars' will produce one bar plot. The plots are generated using ggplot2 and are printed to the current graphical device.
}
\examples{
# Assume 'data.merge' is a data frame with variables 'AGE', 'SEX', 'LIVERMET', 'DIAGTYPE', and a binary variable 'PFSCR'
# binary.barplot(data.merge, c("AGE", "SEX", "LIVERMET", "DIAGTYPE"), "PFSCR")
}
