% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewFunc2.R
\name{rf.imp.barplot}
\alias{rf.imp.barplot}
\title{Variable Importance Plot for Random Forest Model}
\usage{
rf.imp.barplot(model, bar.color)
}
\arguments{
\item{model}{A randomForest model object. This should be the output of a call to randomForest().}

\item{bar.color}{A character string specifying the desired color for the bars in the plot.}
}
\value{
Prints the variable importance scores and generates a bar plot showing the mean decrease in accuracy for each variable in the randomForest model. No return value.
}
\description{
This function calculates and prints the variable importance scores from a Random Forest model based on the mean decrease in accuracy. It also generates a horizontal bar plot visualizing these importance scores, with bars colored as specified by 'bar.color'. The plot is created using the base R plotting functions.
}
\examples{
# Assume 'rf.model' is a trained randomForest model
# rf.imp.barplot(rf.model, bar.color = "blue")
}
