% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outcome_analyses}
\alias{outcome_analyses}
\title{NCT outcome_analyses}
\format{
\subsection{\code{outcome_analyses}}{

A data frame with 256,592 rows and 22 columns. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
}
}
\usage{
outcome_analyses
}
\description{
NCT outcome_analyses data for BIS620
}
\keyword{datasets}
