% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phase_type}
\alias{phase_type}
\title{phase_type data in ctgov.duckdb}
\format{
\subsection{\code{phase_type}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{nct_id}{id of nct}
\item{X, Y, Z}{Accelerometry measurement (in milligravities).}
}
}
}
\usage{
phase_type
}
\description{
Toy data for BIS620
}
\keyword{datasets}
