% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot}
\alias{plot}
\title{Plot number of sex}
\usage{
plot(final)
}
\arguments{
\item{final}{an object inherited from list. It is
assumed to have a \code{adsl} list name}
}
\value{
a plot of number of females and males.
}
\description{
This function Create a plot of number of total females and males
}
\examples{
data(ukb_accel)
summary_sex(ukb_accel)
}
