% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_accel}
\alias{plot_accel}
\title{Plot 3DF Accelerometry Data}
\usage{
plot_accel(x, x_var = "time")
}
\arguments{
\item{x}{a tibble with columns time, X, Y, X}

\item{x_var}{the x-axis variable.}
}
\value{
A ggplot object, faceted by channel, showing the acceleration
or energy (in the Fourier domain) of the data
}
\description{
Plot 3DF Accelerometry Data
}
