% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_project_final.R
\name{plot_adverse_events}
\alias{plot_adverse_events}
\title{Create a stacked barplot of the most frequent adverse events}
\usage{
plot_adverse_events(studies, reported_events)
}
\arguments{
\item{studies}{table/tibble/data.frame, input \code{studies} table filtered by the query
nxd where n is the number of filtered studies and d is the number of features}

\item{reported_events}{table/tibble/data.frame, input \code{reported_events} table
nxd where n is the number of filtered studies and d is the number of features, it is
enough to give as an input \code{reported_events} with the columns \code{nct_id}, \code{organ_system} and \code{event_type}}
}
\description{
ggplot of a stacked barplot of the most frequent adverse events
related to the studies in the input dataset, stacked on the grade (serious and nonserious) of the event;
in the app the dataset is filtered by the query, but the query acts on the table \code{studies} and the
adverse events are stored in the table \code{reported_events}, hence we left join them to filter accordingly
to the query the table \code{reported_events} too
returns void and display a ggplot
NB the histogram is plotted always showing the 10 most frequent adverse events
Note, some studies have no adverse events. Thus,
for such studies nothing be shown on the plot, but the number of such studies with no adverse events
will be shown in the title of the plot
}
