% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adverse_events_genotype.R
\name{plot_adverse_events_genotype}
\alias{plot_adverse_events_genotype}
\title{Create a plot of adverse events by genotype.}
\usage{
plot_adverse_events_genotype(d)
}
\arguments{
\item{d}{A database table containing information about patients, including \code{Genotype} and \code{adverse_events}.}
}
\value{
A bar chart showing the number of adverse events by genotype.
}
\description{
Create a plot of adverse events by genotype.
}
\details{
This function takes a database table of patient data (\code{d}) and generates a bar chart
representing the number of adverse events by genotype. The x-axis corresponds to the genotypes,
and the y-axis represents the count of adverse events in each category.
}
