% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adverse_events_treatment.R
\name{plot_adverse_events_treatment}
\alias{plot_adverse_events_treatment}
\title{Create a plot of adverse events by treatment.}
\usage{
plot_adverse_events_treatment(d)
}
\arguments{
\item{d}{A database table containing information about patients, including \code{ATRT} and \code{adverse_events}.}
}
\value{
A bar chart showing the number of adverse events by treatment.
}
\description{
Create a plot of adverse events by treatment.
}
\details{
This function takes a database table of patient data (\code{d}) and generates a bar chart
representing the number of adverse events by treatment. The x-axis corresponds to the treatments,
and the y-axis represents the count of adverse events in each category.
}
