% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util.R
\name{plot_concurrent_studies}
\alias{plot_concurrent_studies}
\title{Line plot of concurrent studies over time.}
\usage{
plot_concurrent_studies(studies_df)
}
\arguments{
\item{studies_df}{A data frame. The studies to plot the number of concurrent trials for.}
}
\value{
A list with two elements:
\itemize{
\item "concurrent_df": A tibble with a \code{date} column and a \code{count} of the number
of concurrent trials at that date.
\item "concurrent_line_plot": A ggplot object of the line plot of the number of
concurrent studies over time.
}
}
\description{
This function takes a data frame of studies and returns a list
containing a data frame of the number of concurrent studies for each date and
a ggplot object of the line plot of the number of concurrent studies over time.
}
