% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concurrent_trials.R
\name{plot_concurrent_trials}
\alias{plot_concurrent_trials}
\title{Generate a plot of concurrent studies}
\usage{
plot_concurrent_trials(x)
}
\arguments{
\item{x}{The database table.}
}
\value{
A ggplot object showing the concurrent studies.
}
\description{
Generate a plot of concurrent studies
}
