% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_histogram.R
\name{plot_condition_histogram}
\alias{plot_condition_histogram}
\title{#' @title Plot condition histogram}
\usage{
plot_condition_histogram(d)
}
\arguments{
\item{d}{the database table.}
}
\value{
ggplot bar graph showing top 15 conditions
}
\description{
Create a histogram of the conditions returned by a brief
title keyword search
}
