% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_project_final.R
\name{plot_conditions}
\alias{plot_conditions}
\title{Create a histogram of the mot represented conditions related to the different studies}
\usage{
plot_conditions(studies_tibble, conditions_duckdb)
}
\arguments{
\item{studies_tibble}{table/tibble/data.frame, input \code{studies} table filtered by the query
nxd where n is the number of filtered studies and d is the number of features}

\item{conditions_duckdb}{duckDB table, input \code{conditions} table
which is defined at the very beginning of this file. It is
enough to give as an input \code{conditions} table with the columns \code{nct_id} and \code{names} only}
}
\description{
ggplot of an histogram of the conditions related to the studies in the input dataset,
in the app the dataset is filtered by the query, but the query acts on the table \code{studies} and the
conditions are stored in the table \code{conditions}, hence we left join them to filter \code{conditions} table according
to the query too
NB the histogram is plotted showing 10 most frequent conditions in descending order.
The histogram plots less than 10 conditions if there are less than 10 conditions in total after left join
}
