% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conditions_eligibility.R
\name{plot_conditions_eligibility}
\alias{plot_conditions_eligibility}
\title{Plot a histogram of the count of conditions investigated for the selected eligibility criteria}
\usage{
plot_conditions_eligibility(d)
}
\arguments{
\item{d}{the database table.}
}
\description{
Shows a number of conditions studied in a descending order
}
\author{
Jonathan Lee
}
