% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conditions_histogram.R
\name{plot_conditions_histogram}
\alias{plot_conditions_histogram}
\title{The condition histogram of Problem 2}
\usage{
plot_conditions_histogram(x)
}
\arguments{
\item{x}{the database table.}
}
\value{
a histogram of the conditions
}
\description{
Create a histogram of the conditions returned by a brief title keyword search
}
\author{
Min Zhu
}
