% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_countries_frequency.R
\name{plot_countries_frequency}
\alias{plot_countries_frequency}
\title{Create a histogram of the countries that trials in a query are coming from.}
\usage{
plot_countries_frequency(data, num_top_countries)
}
\arguments{
\item{data}{A database table containing information about trials, including \code{country_name}.}

\item{num_top_countries}{The number of top countries to visualize.}
}
\value{
A histogram showing the distribution of countries in the trials.
}
\description{
Create a histogram of the countries that trials in a query are coming from.
}
\details{
This function takes a database table of trials (\code{data}) and generates a histogram
depicting the distribution of countries from which the trials originate. The user can specify
the number of top countries (\code{num_top_countries}) they want to visualize.
}
