% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_countries_map.R
\name{plot_countries_map}
\alias{plot_countries_map}
\title{Map Visualization of Study Count by Country}
\usage{
plot_countries_map(studies)
}
\arguments{
\item{studies}{A database table (tibble) containing the 'country' column with study locations.}
}
\value{
Map plot object.
}
\description{
Produces a choropleth world map indicating the number of studies conducted in each country.
}
