% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_histogram.R
\name{plot_country_histogram}
\alias{plot_country_histogram}
\title{Histogram of number of the first 5 countries that trials in a query are examining.}
\usage{
plot_country_histogram(d)
}
\arguments{
\item{d}{the database table.}
}
\description{
The function gives a histogram showing the countries that trials in a query are examining.
}
