% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{plot_country_map}
\alias{plot_country_map}
\title{Get a country map showing the number of clinical trials per country}
\usage{
plot_country_map(d)
}
\arguments{
\item{d}{data of the studies to get the number of studies trials for.}
}
\value{
a world map with colors showing number of studies trials
}
\description{
Get a country map showing the number of clinical trials per country
}
