% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_project_final.R
\name{plot_detailedness}
\alias{plot_detailedness}
\title{Create scatterplot which shows how detailed/precise the trial is}
\usage{
plot_detailedness(studies_filt, calculated_values)
}
\arguments{
\item{studies_filt}{table/tibble/data.frame, input \code{studies} table filtered by the query
nxd where n is the number of filtered studies and d is the number of features}

\item{calculated_values}{duckDB table, input \code{calculated_values} table
which is defined at the very beginning of this file. It is
enough to give as an input \code{calculated_values} table with the columns \code{nct_id},
\code{number_of_primary_outcomes_to_measure}, \code{number_of_secondary_outcomes_to_measure} and
\code{number_of_other_outcomes_to_measure} only}
}
\description{
scatterplot, which shows the distribution of the trial enrollment (logarithm, base 10),
and the number of total outcomes (sum of primary, secondary and other outomes)
used to measure the results of the trial. The data is also clustered
by the sponsor type;
in the app the dataset is filtered by the query, but the query acts on the table \code{studies}.
This table also contains information about enrollment and sposor_type in the columns \code{enrollment} and
\code{source_class} respectively
However, there is no information about the number of outcomes specified by each trial. This information
is stored in \code{calculated_tables} table,
hence we left join this table to add this information and filter accordingly.
Note, for some studies the information about enrollment or sponsor type is missing. Thus,
such observations will not be shown on the scatterplot, but the number of such observations with missing data
will be shown in the title of the plot
}
