% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facility_map.R
\name{plot_facility_map}
\alias{plot_facility_map}
\title{Plot the world map of the facilities}
\usage{
plot_facility_map(d)
}
\arguments{
\item{d}{the database table.}
}
\description{
Plot a world map with the number of facilities
}
\author{
Anran Jiao
}
