% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facility_status.R
\name{plot_facility_status}
\alias{plot_facility_status}
\title{Plot the status of the facilities}
\usage{
plot_facility_status(d)
}
\arguments{
\item{d}{the database table.}
}
\description{
Plot the number of facilities with different status
}
\author{
Anran Jiao
}
