% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram_uniform_x_axis.R
\name{plot_histogram_uniform_x_axis}
\alias{plot_histogram_uniform_x_axis}
\title{Plot a Histogram for a Column with the x Axis Fixed for All Query Keywords.}
\usage{
plot_histogram_uniform_x_axis(
  x,
  col,
  x_axis,
  x_label,
  title = "",
  na_fill = "NA"
)
}
\arguments{
\item{x}{A tibble data.}

\item{col}{The column used to plot the histogram.}

\item{x_axis}{The value that the \code{x$col} may take.}

\item{x_label}{The label for x axis.}

\item{title}{The title for the histogram. (default "")}

\item{na_fill}{The value to fill NA value. (default "NA")}
}
\value{
A ggplot histogram with itsx axis fixed for all query keywords.
}
\description{
This function plot the histogram for a column in the input database with the x axis fixed for all query keywords.
}
