% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interventions_histogram.R
\name{plot_interventions_histogram}
\alias{plot_interventions_histogram}
\title{Create a histogram of the intervention types that trials in a query are coming from.}
\usage{
plot_interventions_histogram(data)
}
\arguments{
\item{data}{A database table containing information about trials, including \code{intervention_type}.}
}
\value{
A histogram showing the distribution of intervention types in the trials.
}
\description{
Create a histogram of the intervention types that trials in a query are coming from.
}
\details{
This function takes a database table of trials (\code{data}) and generates a histogram
depicting the distribution of intervention types in the trials.
}
