% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Maps the location of the studies on a world map (colored according to number)}
\usage{
plot_map(d)
}
\arguments{
\item{d}{the studies to get the countries for for.}
}
\value{
A histogram showing the countries represented in the study
}
\description{
Maps the location of the studies on a world map (colored according to number)
}
