% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_p_value_histogram.R
\name{plot_p_value_histogram}
\alias{plot_p_value_histogram}
\title{The p value distribution histogram of Feature 6}
\usage{
plot_p_value_histogram(x)
}
\arguments{
\item{x}{the p value table.}
}
\value{
a histogram of p value distribution
}
\description{
Create a histogram of the p value distribution returned by a table of p value
}
\author{
Min Zhu
}
