% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase_histogram.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{Plot phase histogram}
\usage{
plot_phase_histogram(d)
}
\arguments{
\item{d}{the database table.}
}
\value{
ggplot bar graph showing phases
}
\description{
Plots bar chart showing the phases of studies in a query
}
