% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_project_final.R
\name{plot_phase_histogram_new}
\alias{plot_phase_histogram_new}
\title{Create a histogram of the phases of the input studies}
\usage{
plot_phase_histogram_new(x, phases_all)
}
\arguments{
\item{x}{table/tibble/data.frame, input dataset
nxd where n is the number of filtered studies and d is the number of features, one column}

\item{phases_all}{character, it is a vector storing all the possible unique values of "phase"}
}
\description{
ggplot of an histogram of the phases of the studies in the input dataset,
in the app the dataset is filtered by the query;
returning void and displaying a ggplot plot
NB the function always returns the plot with the x-axis uniform wrt the previously performed query on the dataset
}
