% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sponsor_type_overview.R
\name{plot_sponsor_type_overview}
\alias{plot_sponsor_type_overview}
\title{Create a sponsor type overview plot}
\usage{
plot_sponsor_type_overview(x, brief_title_kw)
}
\arguments{
\item{x}{The database table.}

\item{brief_title_kw}{The brief title keywords to look for.}
}
\value{
A ggplot object showing the sponsor type overview.
}
\description{
Create a sponsor type overview plot
}
