% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status_histogram.R
\name{plot_status_histogram}
\alias{plot_status_histogram}
\title{Create a histogram of the status of the study returned by a brief title keyword search}
\usage{
plot_status_histogram(x)
}
\arguments{
\item{x}{the database table.}
}
\description{
Create a histogram of the status of the study returned by a brief title keyword search
}
