% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_status_piechart.R
\name{plot_status_piechart}
\alias{plot_status_piechart}
\title{Pie Chart of Study Status Distribution}
\usage{
plot_status_piechart(studies)
}
\arguments{
\item{studies}{A tibble or dataframe with an 'overall_status' column.}
}
\value{
Pie chart plot object.
}
\description{
Draws a pie chart to represent the distribution of overall statuses of clinical trials.
}
