% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util.R
\name{plot_study_conditions_histogram}
\alias{plot_study_conditions_histogram}
\title{Histogram of the study conditions based on key-word query results}
\usage{
plot_study_conditions_histogram(studies_df, top_n = 8)
}
\arguments{
\item{studies_df}{A data frame. The studies to plot the histogram for.}

\item{top_n}{An integer specifying the number of top conditions to keep.
Default is 8.}
}
\value{
A list with two elements:
\itemize{
\item "summ_df": A data frame of the summarized study conditions.
\item "study_cond_plot": A ggplot object of the histogram of the study conditions.
}
}
\description{
This function takes a data frame of studies and an integer specifying
the number of top conditions to keep, and returns a list containing a data frame
of the summarized study conditions and a ggplot object of the histogram of the
study conditions.
}
