% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_days_genotype.R
\name{plot_survival_days_genotype}
\alias{plot_survival_days_genotype}
\title{Create a plot of survival days by genotype.}
\usage{
plot_survival_days_genotype(d)
}
\arguments{
\item{d}{A database table containing information about patients, including \code{Genotype} and \code{avg_survival_days}.}
}
\value{
A bar chart showing the average number of survival days by genotype.
}
\description{
Create a plot of survival days by genotype.
}
\details{
This function takes a database table of patient data (\code{d}) and generates a bar chart
representing the number of survival days by genotype. The x-axis corresponds to the genotypes,
and the y-axis represents the average number of survival days in each category.
}
