% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_days_treatment.R
\name{plot_survival_days_treatment}
\alias{plot_survival_days_treatment}
\title{Create a plot of survival days by treatment.}
\usage{
plot_survival_days_treatment(d)
}
\arguments{
\item{d}{A database table containing information about patients, including \code{ATRT} and \code{avg_survival_days}.}
}
\value{
A bar chart showing the average number of survival days by treatment.
}
\description{
Create a plot of survival days by treatment.
}
\details{
This function takes a database table of patient data (\code{d}) and generates a bar chart
representing the number of survival days by treatment. The x-axis corresponds to the treatments,
and the y-axis represents the average number of survival days in each category.
}
