% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trial_durations.R
\name{plot_trial_durations}
\alias{plot_trial_durations}
\title{Get the durations of the studies in a query plotted as a histogram}
\usage{
plot_trial_durations(d)
}
\arguments{
\item{d}{the studies to get the durations for for.}
}
\value{
A histogram showing the duration in days of all the studies
}
\description{
Get the durations of the studies in a query plotted as a histogram
}
