% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_project_final.R
\name{plot_world_map}
\alias{plot_world_map}
\title{Create a world map heatmap which shows the number of trials per country}
\usage{
plot_world_map(studies_filt, countries_duckdb, world_map)
}
\arguments{
\item{studies_filt}{table/tibble/data.frame, input \code{studies} table filtered by the query
nxd where n is the number of filtered studies and d is the number of features}

\item{countries_duckdb}{duckDB table, input \code{countries} table
which is defined at the very beginning of this file. It is
enough to give as an input \code{countries} table with the columns \code{nct_id} and \code{name} only}

\item{world_map}{data.frame, input \code{world_map} data.frame, which is defined at
the very beginning of this file. This data.frame is needed to build a world map using geom_map}
}
\description{
heatmap on the world map, which shows the distribution of the studies
from the input dataset across the world;
in the app the dataset is filtered by the query, but the query acts on the table \code{studies}.
However, there is no information about the countries, in
which the sites of trials are located. The information about countries are stored in \code{countries} table,
hence we left join this table to add this information and filter accordingly
Also we need \code{world_map} dataset from the maps package to build the actual world map.
Note, for some studies the information about country is missing in the \code{countries} table. Thus,
such observations will not be shown on the map, but the number of such observations with missing data
will be shown in the title of the plot
}
