% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_and_outcome_data_processing.R
\name{process_intervention_outcomes}
\alias{process_intervention_outcomes}
\title{process intervention and outcome data for plotting}
\usage{
process_intervention_outcomes(get_studies_db)
}
\arguments{
\item{get_studies_db}{database}
}
\description{
This function can aggregate study data and group by outcome_type and intervention_type.
}
