% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_kwds.R
\name{query_kwds}
\alias{query_kwds}
\title{Query Keywords}
\usage{
query_kwds(tbl, kwds, column, ignore_case = TRUE, match_all = FALSE)
}
\arguments{
\item{tbl}{the database table}

\item{kwds}{key words}

\item{column}{the column where searching the key words}

\item{ignore_case}{whether the case will be ignored when searching for the key word. Default is TRUE}

\item{match_all}{whether the selected value shoule match all the key words. Default is FALSE}
}
\value{
a tibble with the selected values
}
\description{
query keywords from database table
}
\details{
Query keywords from a database table.
}
