% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reported_events}
\alias{reported_events}
\title{reported_events data in ctgov.duckdb}
\format{
\subsection{\code{reported_events}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{nct_id}{id of nct}
\item{X, Y, Z}{Accelerometry measurement (in milligravities).}
}
}
}
\usage{
reported_events
}
\description{
Toy data for BIS620
}
\keyword{datasets}
