% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_app.R
\name{run_shiny_app}
\alias{run_shiny_app}
\title{Run Shiny App}
\usage{
run_shiny_app()
}
\arguments{
\item{kwds}{the keywords to look for.}

\item{column}{the column to look for the keywords in.}

\item{ignore_case}{should the case be ignored when searching for a keyword?
(default TRUE)}

\item{match_all}{should we look for values that match all of the keywords
(intersection) or any of the keywords (union)? (default FALSE; union).
Create a histogram of the phases returned by a brief title keyword search}

\item{x}{the database table.
Get the number of concurrent trials for each date in a set of studies}

\item{d}{the studies to get the number of concurrent trials for.}

\item{studies}{the database table.
Create a histogram of the conditions that trials in a query are examining}

\item{num_top_conditions}{the number conditions the user wants to visualize
Create a histogram of the countries that trials in a query are coming from}

\item{num_top_countries}{the number countries the user wants to visualize
Create a histogram of the intervention types that trials in a query are coming from}

\item{data}{the database table.}
}
\value{
No return.

A tibble with a \code{date} column and a \code{count} of the number of
concurrent trials at that date.
Create a plot of the concurrent studies in the query
}
\description{
Starts our Clinical Trials Query Shiny App

Description goes here.
}
