% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run Shiny Application for Clinical Trials Query}
\usage{
runShinyApp()
}
\value{
A running Shiny application on the default web browser.
}
\description{
Run Shiny Application for Clinical Trials Query
}
\details{
The application is composed of a sidebar for input controls and a main panel for displaying the plots and data table output. The following filters can be applied:
\itemize{
\item \code{Brief title keywords}: Filters studies by title keywords.
\item \code{Sponsor Types}: Filters studies by the type of sponsor.
\item \code{Date Range}: Filters studies within a specified date range.
\item \code{Age Range}: Filters studies by participant age range.
\item \code{Lead or Collaborator}: Filters studies by the role of the sponsor.
\item \code{Study Type}: Filters studies by the type of study.
}
The outputs include several plots (phase, concurrent studies, conditions, countries, and status) and a data table displaying the filtered studies.
}
