% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_columns.R
\name{select_colomns}
\alias{select_colomns}
\title{Column Selection}
\usage{
select_colomns(base_data, new_data, colnms)
}
\arguments{
\item{base_data}{the \code{studies} database table.}

\item{new_data}{the new dataset used for feature selection.}

\item{colnms}{the name of selected columns.}
}
\value{
A tibble of new dataset with selected feature.
}
\description{
Select the column in a specific new data set.
}
