% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R
\name{sex}
\alias{sex}
\title{Sex-Based Survival Analysis}
\usage{
sex(dl)
}
\arguments{
\item{dl}{A list containing data frames necessary for creating biodata.}
}
\value{
The function does not return a value but generates multiple survival plots
and prints the results of survival difference analyses for each group.
}
\description{
This function conducts survival analysis based on sex within different biomarker groups.
It generates survival plots and performs survival difference analysis for male and female groups,
as well as for mutant and wild-type biomarker groups.
}
\examples{
# Assuming dl is a list with necessary data frames
data(NCT00364013, package = "bis620.2023")
dl <- NCT00364013
sex(dl)

}
