% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_info.R
\name{show_info}
\alias{show_info}
\title{HTML for single study tab's information page(beyond long description part).}
\usage{
show_info(info, elig)
}
\arguments{
\item{info}{the database table contain the basic information of the studies table.}

\item{elig}{the database table contain the information of the eligibilities table.}
}
\value{
the string for HTML code
}
\description{
The function provides the HTML code that takes tables' variables from users.
}
