% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{spectral_signature}
\alias{spectral_signature}
\alias{spec_sig}
\title{Get the Spectral Signature of Accelerometry Data}
\usage{
spectral_signature(x, take_log = FALSE, inverse = TRUE)
}
\arguments{
\item{x}{an object inherited from a data.frame with columns X, Y, Z, and time
sorted in time.}

\item{take_log}{should the log of the modulus be taken. (Default is \code{TRUE})}

\item{inverse}{should the unnormalized inverse transform is computed.
(Default is \code{TRUE})}
}
\value{
a data frame with the modulus of the Fourier coefficients for the
X, Y, and Z channels.
}
\description{
The spectral signature is calculated by taking the modulus of the
Fourier coefficients of the signal.
}
