% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sponsors}
\alias{sponsors}
\title{Sponsors}
\format{
\subsection{\code{sponsors}}{

A data frame with 761678 rows and 5 columns:
\describe{
\item{id}{A unique identifier}
\item{nct_id}{The NCT id, another identifier}
\item{agency_class}{What type of sponsor is listed}
\item{lead_or_collaborator}{Whether the sponsor was the lead or collaborator for the trial}
\item{name}{Name of the sponsor}
}
}
}
\usage{
sponsors
}
\description{
Taken from ctgov database table, lists sponsors
}
\keyword{datasets}
