% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studies}
\alias{studies}
\title{Studies}
\format{
\subsection{\code{studies}}{

A data frame with 474773 rows and 70 columns:
\describe{
\item{nct_id}{The NCT id, another identifier}
\item{study_first_submitted_date}{Date the study was first submitted}
\item{results_first_submitted_date}{Date the results of the study were first submitted}
...
}
}
}
\usage{
studies
}
\description{
Taken from ctgov database table, lists all the studies
}
\keyword{datasets}
