% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studies_all}
\alias{studies_all}
\title{A NCT studies}
\format{
\subsection{\code{studies_all}}{

A data frame contains full set of studies. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
}
}
\usage{
studies_all
}
\description{
A NCT studies data for BIS620
}
\keyword{datasets}
