% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util.R
\name{summarize_study_conditions}
\alias{summarize_study_conditions}
\title{Summarize Study Conditions}
\usage{
summarize_study_conditions(study_conditions_df, top_n, lump_fct = FALSE)
}
\arguments{
\item{study_conditions_df}{A data frame containing the study conditions data.}

\item{top_n}{An integer specifying the number of top conditions to keep.}

\item{fct_lump}{A logical value indicating whether to apply factor lumping.
Default is FALSE.}
}
\value{
A data frame containing the summarized study conditions data.
}
\description{
This function processes a study conditions data frame and summarizes
the frequency of each condition. It keeps only the top n conditions specified.
If factor lumping is set to TRUE, it applies factor lumping. This is not fully
implemented for the user to enable from the app.
}
