% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{summary_table}
\alias{summary_table}
\title{This function does exploratory analysis of the dataset
A table demonstrating the demographic information is created}
\usage{
summary_table(dl)
}
\arguments{
\item{dl}{a list of tables of the Panitumumab dataset}
}
\value{
a summary table of the demographic information, including age, sex, weight, KRAS type
}
\description{
This function does exploratory analysis of the dataset
A table demonstrating the demographic information is created
}
