% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survATRT_SEX}
\alias{survATRT_SEX}
\title{survival analysis}
\usage{
survATRT_SEX(final)
}
\arguments{
\item{final}{an object inherited from list.}
}
\value{
survival analysis plot and summary statistics
}
\description{
survival analysis
}
