% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivediff.R
\name{survivediff}
\alias{survivediff}
\title{Survival Difference Analysis}
\usage{
survivediff(data, whicharm)
}
\arguments{
\item{data}{A data frame containing survival data, including death day (DTHDY) and death indicator (DTH).}

\item{whicharm}{A character string specifying the column name in \code{data} for stratifying the survival analysis.}
}
\value{
The function prints the results of a survival difference analysis using \code{survdiff} from the 'survival' package.
}
\description{
This function performs a survival difference analysis for different groups in the dataset,
typically different treatment arms, based on a specified stratifying variable.
}
