% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swim_plot.R
\name{swim_plot}
\alias{swim_plot}
\title{A survival plot of death date and actual treatment}
\usage{
swim_plot(data)
}
\arguments{
\item{data}{the data to create survival plot for}
}
\value{
survival plot of death date and actual treatment
}
\description{
A survival plot of death date and actual treatment
}
