% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_adverse_events_treatment.R
\name{test_adverse_events_treatment}
\alias{test_adverse_events_treatment}
\title{Conduct a two-sample t-test to determine whether adverse events differ by treatment.}
\usage{
test_adverse_events_treatment(d)
}
\arguments{
\item{d}{A database table containing information about patients, including \code{ATRT} and \code{adverse_events}.}
}
\value{
A t-test showing whether adverse events differ by treatment.
}
\description{
Conduct a two-sample t-test to determine whether adverse events differ by treatment.
}
\details{
This function takes a database table of patient data (\code{d}) and separates the number of
adverse events by treatment. Then the two resulting groups undergo a two-sample t-test
which will have p<0.05 if their number of adverse events is statistically different.
}
