% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_survival_days_genotype.R
\name{test_survival_days_genotype}
\alias{test_survival_days_genotype}
\title{Conduct a two-sample t-test to determine whether survival days differ by genotype.}
\usage{
test_survival_days_genotype(d)
}
\arguments{
\item{d}{A database table containing information about patients, including \code{Genotype} and \code{DTHDY}.}
}
\value{
A t-test showing whether survival days differ by genotype.
}
\description{
Conduct a two-sample t-test to determine whether survival days differ by genotype.
}
\details{
This function takes a database table of patient data (\code{d}) and separates the number of
survival days by genotype. Then the two resulting groups undergo a two-sample t-test which
will have p<0.05 if their average number of survival days is statistically different.
}
