% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_kw_search.R
\name{title_kw_search}
\alias{title_kw_search}
\title{Query keywords in brief title from a data set}
\usage{
title_kw_search(studies, kw)
}
\arguments{
\item{studies}{the database table}

\item{kw}{key words}
}
\value{
a tibble with the selected values
}
\description{
query keywords in brief title from database table
}
