% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_project_functions.R
\name{transformation_modeling}
\alias{transformation_modeling}
\title{Transforming the data to the variables used in the modelling}
\usage{
transformation_modeling(df)
}
\arguments{
\item{df}{data.frame, dataset to be transformed}
}
\description{
this function transforms the data to create the final variables
which will be used in the further modelling.
}
