% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_hist.R
\name{treat_hist}
\alias{treat_hist}
\title{Histogram of different variables with respect to different treatments}
\usage{
treat_hist(data, count_var, facet_var)
}
\arguments{
\item{data}{the data to create histogram for}

\item{count_var}{the variables to count group treatments for}

\item{facet_var}{the variables to facet group treatments for}
}
\value{
histograms of number of subjects for different treatment
}
\description{
Histogram of different variables with respect to different treatments
}
