% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_shiny_app.R
\name{view_shiny_app}
\alias{view_shiny_app}
\title{Create shiny app to explore clinical trials data}
\usage{
view_shiny_app()
}
\value{
A shiny app visualization with multiple tabs for explortation
}
\description{
Create shiny app to explore clinical trials data
}
