% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cloud.R
\name{word_cloud}
\alias{word_cloud}
\title{Get Word Cloud of Conditions}
\usage{
word_cloud(study, condition)
}
\arguments{
\item{study}{the study table with the number of studies}

\item{condition}{the condition table with the conditions}
}
\value{
a word cloud of key words shown in the name of study
}
\description{
Get Word Cloud of Conditions
}
