% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accel}
\alias{accel}
\title{Accelerometry Data Resampled from UK Biobank}
\format{
\subsection{\code{accel}}{

A data frame with 1,080,000 rows and 3 columns:
\describe{
\item{time}{the time of the measurement}
\item{X, Y, Z}{Accelerometry (in milligravities).}

}
"@source \url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
}
\usage{
accel
}
\description{
Toy Accelerometry Data for BIS620
}
\keyword{datasets}
