% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accel}
\alias{accel}
\title{Accelerometry Data Resampled from UK Biobank}
\format{
\subsection{\code{accel}}{

A data frame with 1,080,000 rows and 4 columns:
\describe{
\item{time}{the time of the measurement}
\item{X, Y, Z}{Accelerometry measurement (in milligravities).}
}
}
}
\usage{
accel
}
\description{
Toy accelerometry data for BIS620
}
\keyword{datasets}
