% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculated.R
\docType{data}
\name{calculated}
\alias{calculated}
\title{Calculated Dataset}
\format{
A data frame with 467704 rows and 3 variables, including:
\describe{
\item{nct_id}{The unique identifier assigned to each clinical study registered on ClinicalTrials.gov}
\item{minimum_age_num}{minimum_age of the patients of a study}
\item{maximum_age_num}{maximum_age of the patients of a study}
}
}
\usage{
calculated
}
\description{
A dataset containing calculated metrics for each study.
}
\keyword{datasets}
