% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calculated_values}
\alias{calculated_values}
\title{NCT calculated_values}
\format{
\subsection{\code{calculated_values}}{

A data frame with 467,212 rows and 19 columns. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
}
}
\usage{
calculated_values
}
\description{
NCT calculated_values data for BIS620
}
\keyword{datasets}
