% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkenvironment.R
\name{check_environment}
\alias{check_environment}
\title{Check Environment}
\usage{
check_environment()
}
\value{
a database connection either connected with local file or remote one.
}
\description{
Check if the required database is available,
otherwise install from remote repository.
}
