% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{NCT conditions}
\format{
\subsection{\code{conditions}}{

A data frame with 808,713 rows and 4 columns. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
}
}
\usage{
conditions
}
\description{
NCT conditions data for BIS620
}
\keyword{datasets}
