% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{NCT Conditions Data}
\format{
\subsection{\code{conditions}}{

A data frame with 808,713 rows and 4 columns. Access the data \href{https://aact.ctti-clinicaltrials.org/pipe_files}{here}.
}
}
\usage{
conditions
}
\description{
NCT Conditions Data
}
\keyword{datasets}
