% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{conditions Data from ctgov}
\format{
\subsection{\code{conditions}}{

A tibble with various information about trials
}
}
\usage{
conditions
}
\description{
conditions data for HW3
}
\keyword{datasets}
