% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Conditions Dataset}
\format{
A data frame with 809607 rows and 4 variables, including:
\describe{
\item{name}{Name of a adverse event}
... and so on ...
}
}
\usage{
conditions
}
\description{
A dataset containing information about the conditions studied.
}
\keyword{datasets}
