% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Conditions Dataset}
\format{
A data frame with 809607 rows and 3 variables, including:
\describe{
\item{nct_id}{The unique identifier assigned to each clinical study registered on ClinicalTrials.gov}
\item{name}{Names for the symptoms}
\item{downcase_name}{Modified names for the symptoms}
}
}
\usage{
conditions
}
\description{
A dataset containing information about the conditions studied.
}
\keyword{datasets}
