% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conditions.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Conditions dataset from ctrialsgov database}
\format{
\subsection{\code{conditions}}{

A data frame with 808,713 rows and 2 columns:
\describe{
\item{nct_id}{The unique identifier assigned to each clinical study registered on ClinicalTrials.gov}
\item{name}{The conditions' names}
}
}
}
\usage{
conditions
}
\description{
Data for BIS620 HW3
}
\keyword{datasets}
