% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{NCT countries}
\format{
\subsection{\code{countries}}{

A data frame with 654,096 rows and 4 columns. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
\describe{
\item{id}{the unique number for each study}
\item{nct_id}{nct id of the study}
\item{name}{country name}
\item{removed}{whether the study is removed}
}
}
}
\usage{
countries
}
\description{
NCT countries data for BIS620
}
\keyword{datasets}
