% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{NCT Countries Data}
\format{
\subsection{\code{countries}}{

A data frame with 654,096 rows and 4 columns. Access the data \href{https://aact.ctti-clinicaltrials.org/pipe_files}{here}.
\describe{
\item{id}{Unique number for each study}
\item{nct_id}{NCT ID of the study}
\item{name}{Country name}
\item{removed}{Whether the study is removed}
}
}
}
\usage{
countries
}
\description{
NCT Countries Data
}
\keyword{datasets}
