% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{countries Data from ctgov}
\format{
\subsection{\code{countries}}{

A tibble with various information about trials
}
}
\usage{
countries
}
\description{
countries data for HW3
}
\keyword{datasets}
