% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Countries Dataset}
\format{
A data frame with 649581 rows and 4 variables, including:
\describe{
\item{nct_id}{xxx}
\item{name}{xxx}
}
}
\usage{
countries
}
\description{
A dataset containing information about the countries involved in the studies.
}
\keyword{datasets}
