% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Countries dataset from ctrialsgov database}
\format{
\subsection{\code{countries}}{

A data frame with 648995 rows and 2 columns:
\describe{
\item{nct_id}{The unique identifier assigned to each clinical study registered on ClinicalTrials.gov}
\item{name}{The name of associated countries}
}
}
}
\usage{
countries
}
\description{
Data for BIS620 HW3
}
\keyword{datasets}
