% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country trials data}
\format{
\subsection{\code{countries}}{

A data frame with 655,026 rows and 4 columns:
\describe{
\item{id}{trial id}
\item{nct_id}{nct_id}
\item{name}{ciuntry name}
\item{removed}{trial removed or not}
}
}
}
\usage{
countries
}
\description{
A subset of data from the trails and countries
}
\keyword{datasets}
