% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_histogram.R
\name{create_condition_histogram}
\alias{create_condition_histogram}
\title{Condition Histogram}
\usage{
create_condition_histogram(x, num_conditions)
}
\arguments{
\item{x}{the database table.}

\item{num_conditions}{Number of most frequent conditions.}
}
\value{
A histogram of the most frequent conditions in a determined amount.
}
\description{
Create a histogram of the conditions.
}
