% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_condition_histogram.R
\name{create_condition_histogram}
\alias{create_condition_histogram}
\title{create_condition_histogram}
\usage{
create_condition_histogram(
  studies,
  conditions,
  sponsor_type,
  status_type,
  brief_title_kw
)
}
\arguments{
\item{studies}{A data frame containing the details of studies.}

\item{conditions}{A dataset of study conditions to join with the studies' data.}

\item{sponsor_type}{A variable representing the sponsor type to filter by. Default is 1.}

\item{status_type}{A variable representing the status type to filter by. Default is 1.}

\item{brief_title_kw}{A keyword to search for in study titles. Default is an empty string.}
}
\value{
A histogram plot showing the distribution of study conditions based on the specified criteria.
}
\description{
This function creates a histogram plot displaying the distribution of study conditions in clinical studies. It allows for optional filtering based on sponsor type, status type, and a keyword search in study titles.
}
\details{
Create a Histogram for Different Conditions
}
