% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_condition_histogram}
\alias{create_condition_histogram}
\title{Histogram of conditions}
\usage{
create_condition_histogram(
  studies,
  sponsors,
  conditions,
  kw,
  sponsortype,
  dates,
  color
)
}
\arguments{
\item{studies}{A tibble. The data table to be filtered}

\item{sponsors}{A tibble. The data table to be joined}

\item{conditions}{A tibble. The data table to be filtered}

\item{kw}{A string. The keyword string entered by the users}

\item{sponsortype}{A string. The sponsortype entered by the users}

\item{dates}{A list of two date values (start and end date)}

\item{color}{A string. The color string chosen by the users}
}
\value{
A plot with filtered entries
}
\description{
Function that creates histogram showing the conditions that trials in a query are examining,
filtered by dates and sponsors information. Apply color feature in the histogram.
}
