% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_intervention_histogram}
\alias{create_intervention_histogram}
\title{Histogram of intervention types}
\usage{
create_intervention_histogram(
  studies,
  sponsors,
  interventions,
  kw,
  interventionType,
  sponsortype,
  dates,
  color
)
}
\arguments{
\item{studies}{A tibble. The data table to be filtered}

\item{sponsors}{A tibble. The data table to be joined}

\item{interventions}{A tibble. The data table to be filtered}

\item{kw}{A string. The keyword string entered by the users}

\item{interventionType}{A string. The intervention type entered by the users}

\item{sponsortype}{A string. The sponsor type entered by the users}

\item{dates}{A list of two date values (start and end date)}

\item{color}{A string. The color string chosen by the users}
}
\value{
A plot with filtered entries
}
\description{
Function that creates a histogram showing the specific intervention that trials in a query are examining,
filtered by intervention type, dates and sponsors information. Apply color feature in the histogram.
}
