% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util.R
\name{create_intervention_pie_data}
\alias{create_intervention_pie_data}
\title{Function that creates a pie chart showing the distribution of intervention type that trials in a query are examining,
filtered by dates and sponsors information.}
\usage{
create_intervention_pie_data(studies, interventions, kw, sponsortype, dates)
}
\arguments{
\item{studies}{A tibble. The data table to be filtered}

\item{kw}{A string. The keyword string entered by the users}

\item{sponsortype}{A string. The sponsor type entered by the users}

\item{dates}{A list of two date values (start and end date)}

\item{intervention}{A tibble. The data table to be filtered}
}
\value{
A piechart with filtered entries
}
\description{
Function that creates a pie chart showing the distribution of intervention type that trials in a query are examining,
filtered by dates and sponsors information.
}
