% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_phase_histogram_plot.R
\name{create_phase_histogram_plot}
\alias{create_phase_histogram_plot}
\title{Phase Histogram}
\usage{
create_phase_histogram_plot(d, way)
}
\arguments{
\item{d}{the database table.}

\item{way}{the filtering choices of allocation ways}
}
\value{
A histogram for numerical phases of clinical trials
}
\description{
Create a histogram of the study phases returned by
Phase Histogram

Create a histogram of the phases.
}
