% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{designs}
\alias{designs}
\title{NCT designs}
\format{
\subsection{\code{designs}}{

A data frame with 462,688 rows and 14 columns. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
}
}
\usage{
designs
}
\description{
NCT designs data for BIS620
}
\keyword{datasets}
