% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concurrent_trials.R
\name{get_concurrent_trials}
\alias{get_concurrent_trials}
\title{Get the number of concurrent trials for each date in a set of studies}
\usage{
get_concurrent_trials(x)
}
\arguments{
\item{x}{The studies to get the number of concurrent trials for.}
}
\value{
A tibble with a \code{date} column and a \code{count} of the number of concurrent trials at that date.
}
\description{
Get the number of concurrent trials for each date in a set of studies
}
