% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util_project_final.R
\name{get_concurrent_trials}
\alias{get_concurrent_trials}
\title{Gets the concurrent trials for each date}
\usage{
get_concurrent_trials(d)
}
\arguments{
\item{d}{tibble/table/dataframe, table containing the studies and the starting and ending dates;
nx2 where n is the number of filtered studies and 2 are the columns corresponding to the start and end date}
}
\description{
Counts the concurrent trials for each date in the dataset then used to plot the concurrent plot;
it returns a tibble \code{all_date} with a \code{date} column and a \code{count} of the number of concurrent trials at that date.
}
