% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facility_info.R
\name{get_facility_info}
\alias{get_facility_info}
\title{The information of facilities, the investigators, and contacts}
\usage{
get_facility_info(facilities, facility_investigators, facility_contacts)
}
\arguments{
\item{facilities}{facilities}

\item{facility_investigators}{facility_investigators}

\item{facility_contacts}{facility_contacts}
}
\value{
A joined table facility_info with investigators of the facilities, and contacts information.
}
\description{
Feature 1: Search the facilities, investigators of the facilities, and contacts information and filter
the data on status, country, role of investigators.
}
\author{
Anran Jiao
}
