% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettermmatrix.R
\name{gettermmatrix}
\alias{gettermmatrix}
\title{Word Frequency for Word Cloud}
\usage{
gettermmatrix(x)
}
\arguments{
\item{x}{the text.}
}
\value{
Frequency of words with descending order.
}
\description{
Get the frequency of words from the text.
}
\details{
Get Term Matrix

Obtain a matrix used for the word cloud display.
}
