% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interventions}
\alias{interventions}
\title{NCT interventions}
\format{
\subsection{\code{interventions}}{

A data frame with 10000 rows and 2 columns. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
}
}
\usage{
interventions
}
\description{
A subset of NCT interventions data for BIS620
}
\keyword{datasets}
