% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interventions}
\alias{interventions}
\title{Subset of NCT interventions data for BIS620.}
\format{
\subsection{\code{interventions}}{

A data frame with 10,000 rows and 2 columns. Access the data \href{https://aact.ctti-clinicaltrials.org/pipe_files}{here}.
}
}
\usage{
interventions
}
\description{
Subset of NCT interventions data for BIS620.
}
\keyword{datasets}
