% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{launchApp}
\alias{launchApp}
\title{Launch a Shiny Application for Clinical Trials Query}
\usage{
launchApp()
}
\description{
This function creates and launches a Shiny application that allows users
to interactively query and visualize clinical trials data. It establishes
a connection to a DuckDB database, reads the data, and sets up a user
interface for data exploration.
}
