% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concurrent_studies.R
\name{plot_concurrent_studies}
\alias{plot_concurrent_studies}
\title{Create a plot of concurrent studies in the query.}
\usage{
plot_concurrent_studies(studies)
}
\arguments{
\item{studies}{A database table containing information about studies, including \code{start_date} and \code{completion_date}.}
}
\value{
A line plot showing the trend of concurrent trials over time.
}
\description{
Create a plot of concurrent studies in the query.
}
\details{
This function takes a database table of studies (\code{studies}) and generates a line plot
representing the number of concurrent trials over time. The x-axis corresponds to the dates,
and the y-axis represents the count of concurrent trials at each date.
}
\seealso{
\code{\link{get_concurrent_trials}}
}
