% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_concurrent_studies.R
\name{plot_concurrent_studies}
\alias{plot_concurrent_studies}
\title{Plot Concurrent Clinical Trials Over Time}
\usage{
plot_concurrent_studies(studies)
}
\arguments{
\item{studies}{the studies to get the number of concurrent trials for.}
}
\value{
Line plot object.
}
\description{
Creates a line plot representing the number of concurrent clinical trials at any given date.
}
