% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_histogram.R
\name{plot_condition_histogram}
\alias{plot_condition_histogram}
\title{Create a histogram of the conditions returned by a brief title keyword search}
\usage{
plot_condition_histogram(x, brief_title_kw)
}
\arguments{
\item{x}{The database table.}

\item{brief_title_kw}{The brief title keywords to look for.}
}
\value{
A ggplot object showing the histogram of conditions.
}
\description{
Create a histogram of the conditions returned by a brief title keyword search
}
