% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_histogram.R
\name{plot_condition_histogram}
\alias{plot_condition_histogram}
\title{Histogram of number of the first 25 conditions that trials in a query are examining.}
\usage{
plot_condition_histogram(d)
}
\arguments{
\item{d}{the database table.}
}
\description{
The function gives a histogram showing the conditions that trials in a query are examining.
}
