% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conditions_histogram.R
\name{plot_conditions_histogram}
\alias{plot_conditions_histogram}
\title{Histogram of Study Conditions}
\usage{
plot_conditions_histogram(studies)
}
\arguments{
\item{studies}{A database table (tibble) containing the study conditions data.}
}
\value{
Histogram plot object.
}
\description{
Creates a histogram displaying the frequency of different conditions in a dataset of
clinical studies.
}
