% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interventions_histogram.R
\name{plot_interventions_histogram}
\alias{plot_interventions_histogram}
\title{Plot Histogram of interventions Types}
\usage{
plot_interventions_histogram(x)
}
\arguments{
\item{x}{A data frame that must contain a column \code{nct_id} for matching with a predefined \code{interventions} dataset, which is assumed to be in the user's environment.}
}
\value{
A histogram. If the input data frame is missing, empty, or lacks intervention types, the function returns a plot with a text annotation indicating the absence of data.
}
\description{
This function takes a dataset and produces two visual representations of interventions types: a histogram. It displays the top 10 interventions types based on their frequency in the dataset.
}
\details{
The function first checks if the input data frame is missing or empty and provides a warning message if so. It then joins the input data frame with the \code{interventions} dataset on \code{nct_id} and calculates the count of each intervention type.

If no intervention types are found, a warning is issued and a plot with a corresponding message is returned. Otherwise, the function creates a histogram showing the count of the top 10 intervention types.
}
