% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Create a world map of study density using the studies dataframe that returned after
filtered by the user}
\usage{
plot_map(x)
}
\arguments{
\item{x}{the dataset to get the number of trails performed in each countries.}
}
\description{
Create a world map of study density using the studies dataframe that returned after
filtered by the user
}
