% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase_histogram.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{Create a histogram of the phases returned by a brief title keyword search.}
\usage{
plot_phase_histogram(x)
}
\arguments{
\item{x}{A database table containing information about trials, including \code{phase}.}
}
\value{
A histogram showing the distribution of phases in the trials.
}
\description{
Create a histogram of the phases returned by a brief title keyword search.
}
\details{
This function takes a database table of trials (\code{x}) and generates a histogram
depicting the distribution of phases in the trials based on a brief title keyword search.
}
