% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase_histogram.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{Create a histogram of the phases returned by a brief title keyword search}
\usage{
plot_phase_histogram(x, brief_title_kw)
}
\arguments{
\item{x}{The database table.}

\item{brief_title_kw}{The brief title keywords to look for.}
}
\value{
A ggplot object showing the histogram of phases.
}
\description{
Create a histogram of the phases returned by a brief title keyword search
}
