% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase_histogram.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{Create Histogram of Study Phases}
\usage{
plot_phase_histogram(x)
}
\arguments{
\item{x}{A database table containing a 'phase' column from which to create the histogram.}
}
\value{
Histogram plot object.
}
\description{
Generates a histogram of different study phases from a dataset.
}
