% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-util.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{Histogram of queried clincial trial phases}
\usage{
plot_phase_histogram(
  studies_df,
  phase_labels = study_phases_u,
  drop_not_applicable = TRUE
)
}
\arguments{
\item{studies_df}{A data frame of queried clinical studies.}

\item{phase_labels}{A character vector specifying the set of phase labels to
show in the plot for standardization. Default is \code{study_phases_u}.}
}
\value{
A list containing the processed data frame and the histogram plot.
}
\description{
This function creates a histogram of study phases. It takes a data
frame of studies and an optional
set of phase labels as input to be used along the x-axis. It processes the data
frame to count the
number of studies in each phase, and then creates a histogram of these counts.
The x-axis of the
histogram represents the study phases, and the y-axis represents the counts.
The function returns
a list containing the processed data frame and the histogram plot.
}
