% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phase_histogram.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{Create a histogram of the phases returned by a brief title keyword search}
\usage{
plot_phase_histogram(x)
}
\arguments{
\item{x}{the database table.}
}
\description{
Create a histogram of the phases returned by a brief title keyword search
}
