% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_hist.R
\name{plot_phase_histogram}
\alias{plot_phase_histogram}
\title{Create a histogram of the phases returned by a brief title keyword search
and sponsor type}
\usage{
plot_phase_histogram(x)
}
\arguments{
\item{x}{the database table.}
}
\value{
a histogram of phases with brief title keyword search
}
\description{
Create a histogram of the phases returned by a brief title keyword search
and sponsor type
}
