% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_keywords.R
\name{query_kwds}
\alias{query_kwds}
\title{Query keywords from a database table.}
\usage{
query_kwds(d, kwds, column, ignore_case = TRUE, match_all = FALSE)
}
\arguments{
\item{d}{A database table.}

\item{kwds}{The keywords to look for in the specified column.}

\item{column}{The column to look for the keywords in.}

\item{ignore_case}{Should the case be ignored when searching for a keyword? (default: TRUE)}

\item{match_all}{Should the function look for values that match all of the keywords (intersection) or any of the keywords (union)? (default: FALSE; union).}
}
\value{
A filtered version of the input database table based on the keyword search.
}
\description{
Query keywords from a database table.
}
\details{
This function performs a keyword search in a specified column of a database table (\code{d}).
It allows for flexible matching options, including case-insensitive search and the choice
between searching for values that match all of the keywords (intersection) or any of the keywords (union).
}
