% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_kwds.R
\name{query_kwds}
\alias{query_kwds}
\title{Query keywords from a database table}
\usage{
query_kwds(d, kwds, column, ignore_case = TRUE, match_all = FALSE)
}
\arguments{
\item{d}{The database table.}

\item{kwds}{The keywords to look for.}

\item{column}{The column to look for the keywords in.}

\item{ignore_case}{Should the case be ignored when searching for a keyword? (default TRUE)}

\item{match_all}{Should we look for values that match all of the keywords (intersection) or any of the keywords (union)? (default FALSE; union).}
}
\value{
A filtered dataset based on keyword search.
}
\description{
Data is queried based on a typed keyword input.
}
