% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_kwds.R
\name{query_kwds}
\alias{query_kwds}
\title{query_kwds}
\usage{
query_kwds(tbl, kwds, column, ignore_case = TRUE, match_all = FALSE)
}
\arguments{
\item{tbl}{A dataset to filter.}

\item{kwds}{Keywords to search for.}

\item{column}{The name of the column in the dataset to perform the search.}

\item{ignore_case}{(Optional) A logical value indicating whether the search should be case-insensitive. Default is TRUE.}

\item{match_all}{(Optional) A logical value indicating whether all keywords must match (AND) or any keyword can match (OR). Default is FALSE.}
}
\value{
A filtered dataset based on the constructed SQL query.
}
\description{
This function construct dplyr query for filtering a dataset based on keyword search criteria applied to a specific column.
}
\details{
Construct  dplyr query for Filtering
}
