% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reported_events}
\alias{reported_events}
\title{reported_events Data from ctgov}
\format{
\subsection{\code{reported_events}}{

A tibble with various information about trials
}
}
\usage{
reported_events
}
\description{
reported_events data for HW3
}
\keyword{datasets}
