% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sponsors}
\alias{sponsors}
\title{NCT sponsors}
\format{
\subsection{\code{sponsors}}{

A data frame with 749,487 rows and 5 columns. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
}
}
\usage{
sponsors
}
\description{
NCT sponsors data for BIS620
}
\keyword{datasets}
