% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sponsors}
\alias{sponsors}
\title{NCT Sponsors Data}
\format{
\subsection{\code{sponsors}}{

A data frame with 749,487 rows and 5 columns. Access the data \href{https://aact.ctti-clinicaltrials.org/pipe_files}{here}.
}
}
\usage{
sponsors
}
\description{
NCT Sponsors Data
}
\keyword{datasets}
