% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sponsors.R
\docType{data}
\name{sponsors}
\alias{sponsors}
\title{Sponsors Dataset}
\format{
A data frame with 750282 rows and 5 variables, including:
\describe{
\item{lead_or_collaborator}{lead or collaborator}
... and so on ...
}
}
\usage{
sponsors
}
\description{
A dataset containing information about the sponsors of the studies.
}
\keyword{datasets}
