% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sponsors.R
\docType{data}
\name{sponsors}
\alias{sponsors}
\title{Sponsors Dataset}
\format{
A data frame with 750282 rows and 4 variables, including:
\describe{
\item{nct_id}{The unique identifier assigned to each clinical study registered on ClinicalTrials.gov}
\item{agency_class}{Class of the sponsor}
\item{lead_or_collaborator}{Whether the sponsor is leading or collaborating}
\item{name}{Name of the sponsor}
}
}
\usage{
sponsors
}
\description{
A dataset containing information about the sponsors of the studies.
}
\keyword{datasets}
