% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sponsors.R
\docType{data}
\name{sponsors}
\alias{sponsors}
\title{Sponsors Dataset}
\format{
A data frame with 750282 rows and 5 variables, including:
\describe{
\item{nct_id}{xxx}
\item{agency_class}{xxx}
\item{lead_or_collaborator}{xxx}
\item{name}{xxx}
}
}
\usage{
sponsors
}
\description{
A dataset containing information about the sponsors of the studies.
}
\keyword{datasets}
