% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_shiny.R
\name{start_shiny}
\alias{start_shiny}
\title{start_shiny}
\usage{
start_shiny()
}
\value{
A shiny app that allows us to explore the contents in the
"ctgov" trials database.
}
\description{
This function starts a shiny app in which we use the "ctgov" trials database
to create visualizations and conduct data analysis.
}
\details{
Call Our Shiny App as a Function
}
