% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studies}
\alias{studies}
\title{NCT studies}
\format{
\subsection{\code{studies}}{

A data frame with 471,252 rows and 70 columns. Please go to the \href{https://aact.ctti-clinicaltrials.org/pipe_files}{clinical trials download site} to get access to the information of this dataset.
}
}
\usage{
studies
}
\description{
NCT studies data for BIS620
}
\keyword{datasets}
