% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studies}
\alias{studies}
\title{Subset of NCT studies data for BIS620.}
\format{
\subsection{\code{studies}}{

A data frame with 10,000 rows and 70 columns. Access the data \href{https://aact.ctti-clinicaltrials.org/pipe_files}{here}.
}
}
\usage{
studies
}
\description{
Subset of NCT studies data for BIS620.
}
\keyword{datasets}
