% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studies}
\alias{studies}
\title{Studies Data from ctgov}
\format{
\subsection{\code{studies}}{

A tibble with various information about trials
}
}
\usage{
studies
}
\description{
Studies data for HW3
}
\keyword{datasets}
