% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\docType{data}
\name{studies}
\alias{studies}
\title{Studies Dataset}
\format{
A data frame with 467704 rows and 70 variables:
\describe{
\item{nct_id}{Unique identifier for a study}
... and so on for the other 70 variables ...
}
}
\usage{
studies
}
\description{
A dataset containing information about various studies. This includes metadata and results of each study.
}
\keyword{datasets}
