% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\docType{data}
\name{studies}
\alias{studies}
\title{Studies Dataset}
\format{
A data frame with 467704 rows and 70 variables:
\describe{
\item{nct_id}{The unique identifier assigned to each clinical study registered on ClinicalTrials.gov}
\item{source_class}{The sponsor type of studies}
\item{overall_status}{The overall status of studies}
\item{phase}{The phase stages of studies}
\item{start_date}{The start date of studies}
\item{study_type}{The study type}
\item{brief_title}{The brif title of studies}
\item{completion_date}{The completion data of studies}
}
}
\usage{
studies
}
\description{
A dataset containing information about various studies. This includes metadata and results of each study.
}
\keyword{datasets}
