% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_studies.R
\docType{data}
\name{studies}
\alias{studies}
\title{Studies dataset from ctrialsgov database}
\format{
\subsection{\code{studies}}{

A data frame with 467212 rows and 8 columns:
\describe{
\item{nct_id}{The unique identifier assigned to each clinical study registered on ClinicalTrials.gov}
\item{source_class}{The sponsor type of studies}
\item{overall_status}{The overall status of studies}
\item{phase}{The phase stages of studies}
\item{start_date}{The start date of studies}
\item{study_type}{The study type}
\item{brief_title}{The brif title of studies}
\item{completion_date}{The completion data of studies}
}
}
}
\usage{
studies
}
\description{
Data for BIS620 HW3
}
\keyword{datasets}
